LPS-102145 Move configuration to web
