--- modules/apps/change-tracking/change-tracking-api/src/main/resources/com/liferay/change/tracking/configuration/packageinfo
@@ -1,1 +1,0 @@
-version 3.0.0
