--- modules/apps/change-tracking/change-tracking-api/src/main/resources/content/Language_ar.properties
@@ -1,2 +1,0 @@
-administrator-role-names=أسماء دور المسؤول
-change-tracking-portal-configuration-name=تعقب التغيير
