--- modules/apps/change-tracking/change-tracking-api/src/main/resources/content/Language_ca.properties
@@ -1,2 +1,0 @@
-administrator-role-names=Noms del rol d'administrador
-change-tracking-portal-configuration-name=Seguiment de canvis
