--- modules/apps/change-tracking/change-tracking-api/src/main/resources/content/Language_de.properties
@@ -1,2 +1,0 @@
-administrator-role-names=Namen von Administratorrollen
-change-tracking-portal-configuration-name=Änderungsverfolgung
