--- modules/apps/change-tracking/change-tracking-api/src/main/resources/content/Language_en.properties
@@ -1,2 +1,0 @@
-administrator-role-names=Administrator Role Names
-change-tracking-portal-configuration-name=Change Tracking
