--- modules/apps/change-tracking/change-tracking-api/src/main/resources/content/Language_es.properties
@@ -1,2 +1,0 @@
-administrator-role-names=Nombres de rol de administrador
-change-tracking-portal-configuration-name=Seguimiento de cambios
