--- modules/apps/change-tracking/change-tracking-api/src/main/resources/content/Language_fi.properties
@@ -1,2 +1,0 @@
-administrator-role-names=Ylläpitäjän Roolinimet
-change-tracking-portal-configuration-name=Muuta Seurantaa
