--- modules/apps/change-tracking/change-tracking-api/src/main/resources/content/Language_fr.properties
@@ -1,2 +1,0 @@
-administrator-role-names=Noms des rôle d'administration
-change-tracking-portal-configuration-name=Modifier le suivi
