--- modules/apps/change-tracking/change-tracking-api/src/main/resources/content/Language_hu.properties
@@ -1,2 +1,0 @@
-administrator-role-names=Rendszergazda szerepkör nevek
-change-tracking-portal-configuration-name=Változások követése
