--- modules/apps/change-tracking/change-tracking-api/src/main/resources/content/Language_ja.properties
@@ -1,2 +1,0 @@
-administrator-role-names=管理者ロール名
-change-tracking-portal-configuration-name=変更の追跡
