--- modules/apps/change-tracking/change-tracking-api/src/main/resources/content/Language_nl.properties
@@ -1,2 +1,0 @@
-administrator-role-names=Namen voor beheerdersrollen
-change-tracking-portal-configuration-name=Wijzigingen bijhouden
