--- modules/apps/change-tracking/change-tracking-api/src/main/resources/content/Language_pt_BR.properties
@@ -1,2 +1,0 @@
-administrator-role-names=Nomes da função de administrador
-change-tracking-portal-configuration-name=Modificar Rastreamento
