--- modules/apps/change-tracking/change-tracking-api/src/main/resources/content/Language_ro.properties
@@ -1,2 +1,0 @@
-administrator-role-names=Administrator Role Names (Automatic Copy)
-change-tracking-portal-configuration-name=Change Tracking (Automatic Copy)
