--- modules/apps/change-tracking/change-tracking-api/src/main/resources/content/Language_sv.properties
@@ -1,2 +1,0 @@
-administrator-role-names=Namn för administratörsroller
-change-tracking-portal-configuration-name=Ändra spårning
