--- modules/apps/change-tracking/change-tracking-api/src/main/resources/content/Language_zh_CN.properties
@@ -1,2 +1,0 @@
-administrator-role-names=管理员角色名称
-change-tracking-portal-configuration-name=更改跟踪
