+++ modules/apps/change-tracking/change-tracking-change-lists-web/src/main/java/com/liferay/change/tracking/change/lists/web/internal/portlet/ChangeListsHistoryPortlet.java
@@ -14,9 +14,9 @@
 
 package com.liferay.change.tracking.change.lists.web.internal.portlet;
 
+import com.liferay.change.tracking.change.lists.web.internal.configuration.CTConfiguration;
 import com.liferay.change.tracking.change.lists.web.internal.constants.CTWebKeys;
 import com.liferay.change.tracking.change.lists.web.internal.display.context.ChangeListsHistoryDisplayContext;
-import com.liferay.change.tracking.configuration.CTConfiguration;
 import com.liferay.change.tracking.constants.CTPortletKeys;
 import com.liferay.change.tracking.service.CTCollectionLocalService;
 import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
@@ -52,7 +52,7 @@ import org.osgi.service.component.annotations.Reference;
  * @author Laszlo Pap
  */
 @Component(
-	configurationPid = "com.liferay.change.tracking.configuration.CTConfiguration",
+	configurationPid = "com.liferay.change.tracking.change.lists.web.internal.configuration.CTConfiguration",
 	configurationPolicy = ConfigurationPolicy.OPTIONAL, immediate = true,
 	property = {
 		"com.liferay.portlet.add-default-resource=false",
