+++ modules/apps/change-tracking/change-tracking-change-lists-web/src/main/resources/content/Language.properties
@@ -1,10 +1,11 @@
 active-change-list=Active Change List
 add-change-list=Add Change List
 adding-x-x-version-x=Adding {0} {1}, version: {2}
+administrator-role-names=Administrator Role Names
 affects=Affects
 an-error-occured-when-trying-publishing-x-x=An error occured when trying to publish '{0}': {1}.
-an-error-occured-when-trying-to-check-x-out=An error occured when trying checkout '{0}' out.
 an-error-occured-when-trying-to-check-x-out-x=An error occured when trying to check '{0}' out: {1}.
+an-error-occured-when-trying-to-check-x-out=An error occured when trying checkout '{0}' out.
 an-error-occured-when-trying-to-delete-x=An error occured when trying delete '{0}'.
 an-error-occured-while-getting-data-from-x=An error occured while getting data from {0}.
 an-error-occured-while-getting-the-active-change-list=An error occured while getting the active change list.
@@ -12,14 +13,15 @@ an-error-occured-while-parsing-data=An error occured while parsing data.
 an-error-occured-while-saving-configuration=An error occured while saving the configuration.
 are-you-sure-you-want-to-delete-x-change-list=Are you sure you want to delete '{0}' change list?
 are-you-sure-you-want-to-publish-x-change-list=Are you sure you want to publish '{0}' change list?
-change-list=Change List
 change-list-description-placeholder=Enter the description of the change list.
-change-list-name=Change List Name
 change-list-name-placeholder=Enter the name of the change list (e.g. Andrew's design changes).
-change-lists=Change Lists
+change-list-name=Change List Name
+change-list=Change List
 change-lists-help=Change lists provide a flexible way to manage different site versions, including its assets and how they are displayed, by using sandboxing concepts. This means users can group changes into one set and publish those changes when they are ready. This is a beta feature and is not supported at this time. Only enable the change lists feature for testing purposes.
 change-lists-history=Change List History
 change-lists-staging-warning=Staging is enabled for at least one site in the current virtual instance. Staging and change lists cannot be used together.
+change-lists=Change Lists
+change-tracking-portal-configuration-name=Change Tracking
 change-type=Change Type
 changes=Changes
 collision-detected-for-x-change-lists=Collision detected for {0} change lists.
